# Improve speed of ZSTD_c_targetCBlockSize