… doc/zstd_manual.html
… programs/benchzstd.c
… programs/benchzstd.h
… programs/zstd.1.md
… programs/zstdcli.c
… tests/Makefile
… tests/fuzz/dictionary_round_trip.c
… tests/fuzzer.c