This PR improves the speed of parameter `ZSTD_c_targetCBlockSize`, which is employed to generate smaller compressed blocks. This parameter was seldom used up to now, and as a consequence, its performance impact was not properly investigated. It proved to be major, as following measurements show.

Some benchmarks for reference, on a local M1 Pro laptop : 
```
/* normal compression */
./zstd -b8 ~/dev/bench/enwik7
 8#enwik7            :  10000000 ->   3184407 (x3.140),   62.8 MB/s, 1267.7 MB/s
 
/* before this PR */
./zstd -b8 ~/dev/bench/enwik7 --target-compressed-block-size=4KB
 8#enwik7            :  10000000 ->   3194310 (x3.131),   3.75 MB/s, 1080.1 MB/s

/* after this PR */
./zstd -b8 ~/dev/bench/enwik7 --target-compressed-block-size=4KB
 8#enwik7            :  10000000 ->   3194256 (x3.131),   62.1 MB/s, 1075.1 MB/s
```

The performance cost of this parameter has been reduced to ~10% of level 1, quickly becoming negligible as a portion of total compression time as level increases (~1% at level 8).

The drawback of this new method is that compressed blocks have a more variable size: they still "target" the requested compressed block size on average, but the variation around the average is higher.
Note that this contract has always been "approximate": even the previous "slow" method would not produce blocks with the requested size, just "close enough". Here, we take full advantage of the nature of this contract for speed benefits.

It is expected that, considering the context in which this parameter is employed, where the goal is to reduce latency to first produced byte, the impact of this variance should not be significant. But the speed improvement, on the other hand, is major, and unlocks usage of this parameter.

If more regularity or guarantees on the compressed size become required or desirable in the future, more refinements will be possible in future updates.